# Mimir query engine: add support for `min_over_time`, `max_over_time`, `sum_over_time` and `avg_over_time`