--- a/pkg/streamingpromql/config.go
+++ b/pkg/streamingpromql/config.go
@@ -21,6 +21,8 @@ type FeatureToggles struct {
 var overTimeFunctionNames = []string{
 	"count_over_time",
 	"last_over_time",
+	"max_over_time",
+	"min_over_time",
 	"present_over_time",
 }