--- a/pkg/streamingpromql/testdata/ours/functions.test
+++ b/pkg/streamingpromql/testdata/ours/functions.test
@@ -67,22 +67,22 @@ eval range from 0 to 6m step 1m floor(some_metric)
 clear
 
 load 1m
-  some_metric 0 1 2 3 _ _ {{schema:3 sum:4 count:4 buckets:[1 2 1]}} {{schema:3 sum:4 count:4 buckets:[1 2 1]}}
+  some_metric{foo="bar"} 0 1 2 3 _ _ {{schema:3 sum:4 count:4 buckets:[1 2 1]}} {{schema:3 sum:4 count:4 buckets:[1 2 1]}}
 
 eval range from 0 to 7m step 1m count_over_time(some_metric[3m])
-  {} 1 2 3 4 3 2 2 2
+  {foo="bar"} 1 2 3 4 3 2 2 2
 
 eval range from 0 to 7m step 1m count_over_time(some_metric[5s])
-  {} 1 1 1 1 _ _ 1 1
+  {foo="bar"} 1 1 1 1 _ _ 1 1
 
 eval range from 0 to 7m step 1m last_over_time(some_metric[3m])
-  some_metric 0 1 2 3 3 3 {{schema:3 sum:4 count:4 buckets:[1 2 1]}} {{schema:3 sum:4 count:4 buckets:[1 2 1]}}
+  some_metric{foo="bar"} 0 1 2 3 3 3 {{schema:3 sum:4 count:4 buckets:[1 2 1]}} {{schema:3 sum:4 count:4 buckets:[1 2 1]}}
 
 eval range from 0 to 7m step 1m last_over_time(some_metric[5s])
-  some_metric 0 1 2 3 _ _ {{schema:3 sum:4 count:4 buckets:[1 2 1]}} {{schema:3 sum:4 count:4 buckets:[1 2 1]}}
+  some_metric{foo="bar"} 0 1 2 3 _ _ {{schema:3 sum:4 count:4 buckets:[1 2 1]}} {{schema:3 sum:4 count:4 buckets:[1 2 1]}}
 
 eval range from 0 to 7m step 1m present_over_time(some_metric[3m])
-  {} 1 1 1 1 1 1 1 1
+  {foo="bar"} 1 1 1 1 1 1 1 1
 
 eval range from 0 to 7m step 1m present_over_time(some_metric[5s])
-  {} 1 1 1 1 _ _ 1 1
+  {foo="bar"} 1 1 1 1 _ _ 1 1