--- a/pkg/streamingpromql/testdata/ours/functions.test
+++ b/pkg/streamingpromql/testdata/ours/functions.test
@@ -86,3 +86,15 @@ eval range from 0 to 7m step 1m present_over_time(some_metric[3m])
 
 eval range from 0 to 7m step 1m present_over_time(some_metric[5s])
   {foo="bar"} 1 1 1 1 _ _ 1 1
+
+eval range from 0 to 7m step 1m min_over_time(some_metric[3m])
+  {foo="bar"} 0 0 0 0 1 2 3 _
+
+eval range from 0 to 7m step 1m min_over_time(some_metric[5s])
+  {foo="bar"} 0 1 2 3 _ _ _ _
+
+eval range from 0 to 7m step 1m max_over_time(some_metric[3m])
+  {foo="bar"} 0 1 2 3 3 3 3 _
+
+eval range from 0 to 7m step 1m max_over_time(some_metric[5s])
+  {foo="bar"} 0 1 2 3 _ _ _ _