--- a/pkg/streamingpromql/testdata/ours/functions.test
+++ b/pkg/streamingpromql/testdata/ours/functions.test
@@ -69,6 +69,7 @@ clear
 load 1m
   some_metric{foo="bar"} 0 1 2 3 _ _ {{schema:3 sum:4 count:4 buckets:[1 2 1]}} {{schema:3 sum:5 count:3 buckets:[2 5 4]}}
   some_nhcb_metric{baz="bar"} {{schema:-53 sum:1 count:5 custom_values:[5 10] buckets:[1 4]}} {{schema:-53 sum:15 count:2 custom_values:[5 10] buckets:[0 2]}} {{schema:-53 sum:3 count:15 custom_values:[5 10] buckets:[7 8]}}
+  some_inf_and_nan_metric{foo="baz"} 0 1 2 3 Inf Inf Inf NaN NaN NaN NaN 8 7 6
 
 eval range from 0 to 7m step 1m count_over_time(some_metric[3m])
   {foo="bar"} 1 2 3 4 3 2 2 2
@@ -94,11 +95,17 @@ eval range from 0 to 7m step 1m min_over_time(some_metric[3m])
 eval range from 0 to 7m step 1m min_over_time(some_metric[5s])
   {foo="bar"} 0 1 2 3 _ _ _ _
 
+eval range from 0 to 16m step 1m min_over_time(some_inf_and_nan_metric[3m])
+  {foo="baz"} 0 0 0 0 1 2 3 Inf Inf Inf NaN 8 7 6 6 6 6
+
 eval range from 0 to 7m step 1m max_over_time(some_metric[3m])
   {foo="bar"} 0 1 2 3 3 3 3 _
 
 eval range from 0 to 7m step 1m max_over_time(some_metric[5s])
   {foo="bar"} 0 1 2 3 _ _ _ _
+
+eval range from 0 to 16m step 1m max_over_time(some_inf_and_nan_metric[3m])
+  {foo="baz"} 0 1 2 3 Inf Inf Inf Inf Inf Inf NaN 8 8 8 8 7 6
 
 eval_warn range from 0 to 10m step 1m sum_over_time(some_metric[3m])
   {foo="bar"} 0 1 3 6 6 5 _ {{schema:3 sum:9 count:7 buckets:[3 7 5]}} {{schema:3 sum:9 count:7 buckets:[3 7 5]}} {{schema:3 sum:9 count:7 buckets:[3 7 5]}} {{schema:3 sum:5 count:3 buckets:[2 5 4]}}
@@ -115,6 +122,9 @@ eval range from 0 to 7m step 1m sum_over_time(some_metric[5s])
 eval range from 0 to 2m step 1m sum_over_time(some_nhcb_metric[3m])
   {baz="bar"} {{schema:-53 sum:1 count:5 custom_values:[5 10] buckets:[1 4]}} {{schema:-53 sum:16 count:7 custom_values:[5 10] buckets:[1 6]}} {{schema:-53 sum:19 count:22 custom_values:[5 10] buckets:[8 14]}}
 
+eval range from 0 to 16m step 1m sum_over_time(some_inf_and_nan_metric[3m])
+  {foo="baz"} 0 1 3 6 Inf Inf Inf NaN NaN NaN NaN NaN NaN NaN 21 13 6
+
 eval_warn range from 0 to 10m step 1m avg_over_time(some_metric[3m])
   {foo="bar"} 0 0.5 1 1.5 2 2.5 _ {{schema:3 sum:4.5 count:3.5 buckets:[1.5 3.5 2.5]}} {{schema:3 sum:4.5 count:3.5 buckets:[1.5 3.5 2.5]}} {{schema:3 sum:4.5 count:3.5 buckets:[1.5 3.5 2.5]}} {{schema:3 sum:5 count:3 buckets:[2 5 4]}}
 
@@ -129,3 +139,6 @@ eval range from 0 to 7m step 1m avg_over_time(some_metric[5s])
 
 eval range from 0 to 2m step 1m avg_over_time(some_nhcb_metric[3m])
   {baz="bar"} {{schema:-53 sum:1 count:5 custom_values:[5 10] buckets:[1 4]}} {{schema:-53 sum:8 count:3.5 custom_values:[5 10] buckets:[0.5 3]}} {{schema:-53 sum:6.333333333333334 count:7.333333333333333 custom_values:[5 10] buckets:[2.666666666666667 4.666666666666666]}}
+
+eval range from 0 to 16m step 1m avg_over_time(some_inf_and_nan_metric[3m])
+  {foo="baz"} 0 0.5 1 1.5 Inf Inf Inf NaN NaN NaN NaN NaN NaN NaN 7 6.5 6