--- a/pkg/streamingpromql/testdata/upstream/at_modifier.test
+++ b/pkg/streamingpromql/testdata/upstream/at_modifier.test
@@ -84,14 +84,12 @@ eval instant at 10s metric @ 1m40s
 #   {job="2"} -20
 
 # Millisecond precision.
-# Unsupported by streaming engine.
-# eval instant at 100s metric_ms @ 1.234
-#   metric_ms 1234
+eval instant at 100s metric_ms @ 1.234
+  metric_ms 1234
 
 # Range vector selectors.
-# Unsupported by streaming engine.
-# eval instant at 25s sum_over_time(metric{job="1"}[100s] @ 100)
-#   {job="1"} 55
+eval instant at 25s sum_over_time(metric{job="1"}[100s] @ 100)
+  {job="1"} 55
 
 # Unsupported by streaming engine.
 # eval instant at 25s sum_over_time(metric{job="1"}[100s] @ 100 offset 50s)