… CHANGELOG.md
… pkg/streamingpromql/benchmarks/benchmarks.go
… pkg/streamingpromql/engine_concurrency_test.go
… pkg/streamingpromql/engine_test.go
… pkg/streamingpromql/floats/kahan_test.go
… pkg/streamingpromql/operators/aggregation_test.go
… pkg/streamingpromql/testdata/upstream/functions.test
… pkg/streamingpromql/testdata/upstream/range_queries.test